Split `HttpClient` into simpler structs
`HttpFetcher` is responsible for doing HTTP requests and handling
redirects. `HttpClient` is a wrapper over `HttpFetcher` that takes
robots.txt into account.
